requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.6
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.10
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
sqlparse==0.4.3
termcolor==2.0.1
tifffile==2022.10.10
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.9.0
